refactor: remove hardware hooks
