+++ packages/kit/src/components/Hardware/index.tsx
@@ -1,295 +1,153 @@
-import { useForm } from 'react-hook-form';
-import { StyleSheet } from 'react-native';
-import { useMedia } from 'tamagui';
+import { useCallback } from 'react';
+
+import { Dialog } from '@onekeyhq/components';
 
 import {
-  Alert,
-  Button,
-  Dialog,
-  Form,
-  IconButton,
-  Input,
-  LottieView,
-  SizableText,
-  Stack,
-  Toast,
-  XStack,
-  useDialogInstance,
-} from '@onekeyhq/components';
-import type { IButtonProps } from '@onekeyhq/components';
+  ConfirmOnClassic,
+  ConfirmPassphrase,
+  EnterPassphraseOnDevice,
+  EnterPhase,
+  EnterPin,
+  EnterPinOnDevice,
+} from './Hardware';
 
-export function ConfirmOnClassic() {
-  return (
-    <XStack alignItems="center" mb="$-2.5" mt="$2.5" ml="$-2.5">
-      <Stack
-        w="$16"
-        h="$16"
-        bg="$bgStrong"
-        borderRadius="$2"
-        style={{ borderCurve: 'continuous' }}
-        justifyContent="center"
-        alignItems="center"
-      >
-        <LottieView
-          width={64}
-          height={64}
-          source={require('../../../assets/animations/confirm-on-classic.json')}
-        />
-      </Stack>
-      <SizableText size="$bodyLgMedium" textAlign="center" pl="$4">
-        Confirm on Device
-      </SizableText>
-    </XStack>
-  );
-}
+const mockListenDeviceResult = () => {
+  const actions: (() => void | undefined)[] = [];
+  return {
+    run: () =>
+      new Promise<void>((resolve, reject) => {
+        actions[0] = resolve;
+        actions[1] = reject;
+      }),
+    confirm: () => {
+      actions?.[0]();
+    },
+    cancel: () => {
+      actions?.[1]();
+    },
+  };
+};
 
-export function EnterPin({
-  onConfirm,
-  switchOnDevice,
-}: {
-  onConfirm: () => void;
-  switchOnDevice: () => void;
-}) {
-  return (
-    <Stack>
-      <Stack
-        borderWidth={StyleSheet.hairlineWidth}
-        borderColor="$borderSubdued"
-        borderRadius="$2"
-        overflow="hidden"
-        style={{
-          borderCurve: 'continuous',
-        }}
-      >
-        <XStack
-          h="$12"
-          alignItems="center"
-          px="$3"
-          borderBottomWidth={StyleSheet.hairlineWidth}
-          borderColor="$borderSubdued"
-          bg="$bgSubdued"
-        >
-          <SizableText pl="$6" textAlign="center" flex={1} size="$heading4xl">
-            ••••••
-          </SizableText>
-          <IconButton variant="tertiary" icon="XBackspaceOutline" />
-        </XStack>
-        <XStack flexWrap="wrap">
-          {Array.from({ length: 9 }).map((_, index) => (
-            <Stack
-              key={index}
-              flexBasis="33.3333%"
-              h="$14"
-              borderRightWidth={StyleSheet.hairlineWidth}
-              borderBottomWidth={StyleSheet.hairlineWidth}
-              borderColor="$borderSubdued"
-              justifyContent="center"
-              alignItems="center"
-              {...((index === 2 || index === 5 || index === 8) && {
-                borderRightWidth: 0,
-              })}
-              {...((index === 6 || index === 7 || index === 8) && {
-                borderBottomWidth: 0,
-              })}
-              hoverStyle={{
-                bg: '$bgHover',
-              }}
-              pressStyle={{
-                bg: '$bgActive',
-              }}
-              focusable
-              focusStyle={{
-                outlineColor: '$focusRing',
-                outlineOffset: -2,
-                outlineWidth: 2,
-                outlineStyle: 'solid',
-              }}
-            >
-              <Stack w="$2.5" h="$2.5" borderRadius="$full" bg="$text" />
-            </Stack>
-          ))}
-        </XStack>
-      </Stack>
-      {/* TODO: add loading state while waiting for result */}
-      <Button
-        mt="$5"
-        $md={
-          {
-            size: 'large',
-          } as IButtonProps
-        }
-        variant="primary"
-        onPress={() => {
-          Toast.error({
-            title: 'Wrong PIN',
-          });
-          onConfirm();
-        }}
-      >
-        Confirm
-      </Button>
-      <Button
-        m="$0"
-        mt="$2"
-        $md={
-          {
-            size: 'large',
-          } as IButtonProps
-        }
-        variant="tertiary"
-        onPress={switchOnDevice}
-      >
-        Enter on Device
-      </Button>
-    </Stack>
-  );
-}
+export const confirmOnClassic = async () => {
+  const event = mockListenDeviceResult();
+  const dialog = Dialog.show({
+    disableDrag: true,
+    dismissOnOverlayPress: false,
+    showFooter: false,
+    floatingPanelProps: {
+      mt: '$5',
+      mb: 'auto',
+    },
+    renderContent: <ConfirmOnClassic />,
+    onDismiss: () => {
+      event.cancel();
+    },
+  });
+  setTimeout(async () => {
+    event.confirm();
+    await dialog.close();
+  }, 3500);
+  await event.run();
+};
 
-export function EnterPinOnDevice() {
-  return (
-    // height must be specified on Sheet View.
-    <Stack borderRadius="$3" bg="$bgSubdued" height={230}>
-      <LottieView
-        width="100%"
-        height="100%"
-        source={require('../../../assets/animations/enter-pin-on-classic.json')}
-      />
-    </Stack>
-  );
-}
+export const confirmPinOnDevice = async () => {
+  const event = mockListenDeviceResult();
+  const dialog = Dialog.show({
+    title: 'Enter PIN on Device',
+    showFooter: false,
+    renderContent: <EnterPinOnDevice />,
+    onCancel: () => {
+      event.cancel();
+    },
+  });
+  setTimeout(async () => {
+    event.confirm();
+    await dialog.close();
+  }, 3500);
+  await event.run();
+};
 
-export function EnterPhase({
-  onConfirm,
-  switchOnDevice,
-}: {
-  onConfirm: () => void;
-  switchOnDevice: () => void;
-}) {
-  const form = useForm();
-  const media = useMedia();
-  const dialog = useDialogInstance();
-  return (
-    <Stack>
-      <Stack pb="$5">
-        <Alert
-          title="Protect Your Passphrase: Irrecoverable if Lost."
-          type="warning"
-        />
-      </Stack>
-      <Form form={form}>
-        <Form.Field name="passphrase" label="Passphrase">
-          <Input
-            placeholder="Enter passphrase"
-            {...(media.md && {
-              size: 'large',
-            })}
-          />
-        </Form.Field>
-        <Form.Field name="confirmPassphrase" label="Confirm Passphrase">
-          <Input
-            placeholder="Re-enter your passphrase"
-            {...(media.md && {
-              size: 'large',
-            })}
-          />
-        </Form.Field>
-      </Form>
-      {/* TODO: add loading state while waiting for result */}
-      <Button
-        mt="$5"
-        $md={
-          {
-            size: 'large',
-          } as IButtonProps
-        }
-        variant="primary"
-        onPress={async () => {
+export const confirmByPin = async () => {
+  const event = mockListenDeviceResult();
+  const dialog = Dialog.show({
+    title: 'Enter PIN',
+    showFooter: false,
+    renderContent: (
+      <EnterPin
+        onConfirm={async () => {
+          event.confirm();
           await dialog.close();
-          Dialog.show({
-            icon: 'CheckboxSolid',
-            title: 'Keep Your Wallet Accessible?',
-            description:
-              'Save this wallet to your device to maintain access after the app is closed. Unsaved wallets will be removed automatically.',
-            onConfirm: () => {
-              onConfirm();
-            },
-            onConfirmText: 'Save Wallet',
-            confirmButtonProps: {
-              variant: 'secondary',
-            },
-            onCancel: () => console.log('canceled'),
-            onCancelText: "Don't Save",
-          });
         }}
-      >
-        Confirm
-      </Button>
-      <Button
-        m="$0"
-        mt="$2"
-        $md={
-          {
-            size: 'large',
-          } as IButtonProps
-        }
-        variant="tertiary"
-        onPress={switchOnDevice}
-      >
-        Enter on Device
-      </Button>
-    </Stack>
-  );
-}
-
-export function EnterPassphraseOnDevice() {
-  return (
-    <Stack borderRadius="$3" bg="$bgSubdued" height={230}>
-      <LottieView
-        width="100%"
-        height="100%"
-        source={require('../../../assets/animations/enter-passphrase-on-classic.json')}
+        switchOnDevice={async () => {
+          await dialog.close();
+          await confirmPinOnDevice();
+        }}
       />
-    </Stack>
-  );
-}
+    ),
+    onDismiss: () => {
+      event.cancel();
+    },
+  });
+  await event.run();
+};
 
-export function ConfirmPassphrase({
-  onConfirm,
-  switchOnDevice,
-}: {
-  onConfirm: () => void;
-  switchOnDevice: () => void;
-}) {
-  return (
-    <Stack>
-      {/* TODO: switch size to large when media.md */}
-      <Input placeholder="Enter your passphrase" />
-      {/* TODO: add loading state while waiting for result */}
-      <Button
-        mt="$5"
-        $md={
-          {
-            size: 'large',
-          } as IButtonProps
-        }
-        variant="primary"
-        onPress={onConfirm}
-      >
-        Confirm
-      </Button>
-      <Button
-        m="$0"
-        mt="$2"
-        $md={
-          {
-            size: 'large',
-          } as IButtonProps
-        }
-        variant="tertiary"
-        onPress={switchOnDevice}
-      >
-        Enter on Device
-      </Button>
-    </Stack>
-  );
-}
+export const confirmPhraseOnDevice = async () => {
+  const event = mockListenDeviceResult();
+  Dialog.show({
+    title: 'Enter Passphrase on Device',
+    showFooter: false,
+    renderContent: <EnterPassphraseOnDevice />,
+    onCancel: () => {
+      event.cancel();
+    },
+  });
+  await event.run();
+};
+
+export const confirmPhrase = async () => {
+  const event = mockListenDeviceResult();
+  const dialog = Dialog.show({
+    title: 'Enter Passphrase',
+    showFooter: false,
+    renderContent: (
+      <EnterPhase
+        onConfirm={async () => {
+          event.confirm();
+          await dialog.close();
+        }}
+        switchOnDevice={async () => {
+          await dialog.close();
+          await confirmPhraseOnDevice();
+        }}
+      />
+    ),
+    onCancel: () => {
+      event.cancel();
+    },
+  });
+  await event.run();
+};
+
+export const confirmPassphrase = async () => {
+  const event = mockListenDeviceResult();
+  const dialog = Dialog.show({
+    title: 'Confirm Passphrase',
+    showFooter: false,
+    renderContent: (
+      <ConfirmPassphrase
+        onConfirm={async () => {
+          event.confirm();
+          await dialog.close();
+        }}
+        switchOnDevice={async () => {
+          await dialog.close();
+          await confirmPhraseOnDevice();
+        }}
+      />
+    ),
+    onCancel: () => {
+      event.cancel();
+    },
+  });
+  await event.run();
+};
